[IMP] pivot panel: make cog wheel menu use `Menu`
The cog wheel menu in the pivot panel was using a custom implementation
of a menu, because we didn't want it to be as large as a regular menu.

This commit changes the cog wheel menu to use the `Menu` component, and
add a new props to `Menu` to set its width.

closes odoo/o-spreadsheet#4698

Task: 4008880
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>