[FIX] composer: layout break when composer assistant is opened
When typing =S in the composer, the composer assistant will open.
If there wasn't enough space on right (default 300px width), this would
break the layout. This PR addresses the issue by adding a width check
for the composer assistant. If the width condition is met, the style

closes odoo/o-spreadsheet#4719

Right: 0px is applied to prevent the layout from breaking.
Task: 4037213
X-original-commit: 52d2cd36012c581b4f3d0ca325286fb4eb61d67c
Signed-off-by: Dhrutik Patel (dhrp) <dhrp@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>