[FIX] composer: layout breaks when composer assistant is opened
In FW, a PR#4721 was mistakenly approved (r+) because it had no
conflicts, but it did not resolve the intended issue in the master
branch. This new PR addresses that issue.

Description:

When typing =S in the composer, the composer assistant opens. If
there isn't enough space on the right (default 300px width),
it breaks the layout.

This PR fixes the issue by adding a width check for the composer
assistant. If the width condition is met, the style right: 0px is
applied to prevent the layout from breaking.

closes odoo/o-spreadsheet#4723

Task: 4037213
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>