[FIX] charts: Add missing "showValue" handler for Pie and waterfall charts
The typing of chart.js is NOT our friend and it shows again. Expecially
the type `ChartOptions` that applies a deepPartial on each of its
properties, including the plugin recently introduced: `chartShowValuesPlugin`.

Long story short, if provided with `showValues=true`, the plugin
REQUIRES a callback function to format the data value. Unfortunately,
the `DeepPartial` completely hides this requirement.

This bug was also missed because we do not directly test chart.js lib in
our tests and it will be the subject of a task in the near future. In
the meantime, we will strengthen the test case in our main integration
(Odoo).

closes odoo/o-spreadsheet#4788

Task: 4081436
X-original-commit: da06107e8f2812e8ecb2aac92d58348f090d95ee
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>