[FIX] sheetview: Avoid useless cleanViewport
cleanViewports is called at every command handling in `SheetViewPlugin`
but it only makes sense do to so for commands that could actually impact
the internal viewport mapping of the plugin. Namely,
UNDO/REDO/DELETE_SHEET commands could have an impact on that mapping.

This revision limits the calls to cleanViewports for those specific
commands.

When deleting a row on a sheet of 26*10_000 cells, this spares around
100ms of computation.

closes odoo/o-spreadsheet#4935

Task: 4084860
X-original-commit: 2f972679aa7e98373e90b480b3f7d660b9ea615a
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>