[PERF] SheetPlugin: Faster header addition/deletion
When adding/deleting a header, we have to shift all the cells
accordingly. This is currently done by the `SheetPlugin` which
dispatches the command "UPDATE_CELL_POSITION". However, it is
the only plugin that handles the said command.
In certain circumstances, the overhead of the whole `dispatch` process
can become very problematic as the number of subcommands increases.

e.g. in a sheet with 20000 rows, 20 columns and we delete its
first row, we will dispatch 399980 commands `UPDATE_CELL_POSITION`
whereas we could simply call the internal command handler of the
plugin and bypass the overhead.

With this revision, we bypass the subdispatch and the removal of
the first row drops from 850 ms to 230 ms so ~73% performance improvement.

The issue is also present for the addition of headers.
Adding a row before the first one in a 20000*20 sheet drops from
962ms to 360ms so ~62.5% perf improvement.

Task: 4084858
Part-of: odoo/o-spreadsheet#4749
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>