[IMP] pivot: allow to format a measure
This commit allows to define a custom format on pivot measures by
selecting any cell with a measure value and applying a format.
The format applies to the measure, which means to all cells being
values of that measure, even those that were not selected when applying
the format.
This behavior is chosen because it's the most probable user intention when
he wants to format pivot measure cells.

It's particularly useful for calculated measures where the automatic
format inferred from the formula might not be what the users intented.

Task: 3956676
Part-of: odoo/o-spreadsheet#4657
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>