[IMP] format: handled more complex text format
Before, the only text format that was handled was the simple `@` format.
Now that we handle more complex text formats, we should make sure
they are handled correctly during the cell creation.

Also removed the `PLAIN_TEXT_FORMAT` variable. The variable made it
looks like the `@` was an immutable special case, whereas `@`  is just
a text placeholder that can be used in more complex text formats.

Task: 4083126
Part-of: odoo/o-spreadsheet#4769
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>