[FIX] table: fix bug when edit table with invalid range
Before this commit, when editing a table and entering an invalid range, the entire spreadsheet stops working and causes the browser to crash. This is due to an infinite call to `onWillUpdateProps` of the `SelectionInput` component.

This commit fixes the problem by removing a useless range validity check when the input changes.

closes odoo/o-spreadsheet#4868

Task: 4102425
X-original-commit: 6e6a23ff7528e9c290031e14c88f905c64b65732
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Mehdi Rachico (mera) <mera@odoo.com>