[FIX] charts: fix combo chart export
Task Description

Since the chart customization PR, there is an issue in the combo chart
export where there is a mismatch between the vertical left and right
axes, leading to an wrong axis definition when exporting a combo chart
and a removed chart when loading it to excel.
This commits aims to fix this issue.

Related Task

closes odoo/o-spreadsheet#5018

Task: 4092078
X-original-commit: fbdcb2bcb5f89ed432e857e556fad25a6e635f8e
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Hendrickx Anthony (anhe) <anhe@odoo.com>