[IMP] Composer: Implement range selection/edition toggle
Currently, the only way to select a range in a formula is to write some
text such that the next argument could be a range (e.g. write '=sum(').
If you edit an existing formula, you need to delete a range argument in
order to 'activate' the range selection mode. While you can use the
graphical tool to move the range around, it can be bothersome to find
which range is where and then having to move it around.

This revision adds the possibility to switch to a range selection mode
when the cursor is on a reference by pressing F2. This allows the user
to click on a range in the composer, switch to 'selecting' mode and then
cllick on another range to replace the current one.
Pressing the shortcut again will switch back to 'editing' mode which can
be helpful to quickly edit formulas.

closes odoo/o-spreadsheet#5012

Task: 4105364
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>