[FIX] find & replace: remove debounce from store
The method `updateSearchContent` was debounced inside the f&r store.
This does not work as expected, as the re-rendering of the panel was
done when calling the `updateSearchContent` method, and not when the
debounced function was called and the store state was updated.

I'm not sure why it still worked 50% of the time.

closes odoo/o-spreadsheet#4847

Task: 4102172
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>