[IMP] composer: collapse function assistant details
This commit it a part of a series of changes to the composer auto-complete
and formula assistant.

The goal is to show them both. Currently the auto-complete has priority
over the assistant which is not displayed when the auto-complete is displayed.
Which means you can't know which parameter is which.

This commit allows to collapse the function details from the
formula assistant. It allows to make room for the auto-complete.

Note that the collapsible behavior cannot be tested with jest because
it relies on bootstrap which is not included when running tests.

Task: 3956676
Part-of: odoo/o-spreadsheet#4657
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>