[IMP] format: add the handling of repeated character
This commit adds the handling of the `*` asterisk character in the format,
which repeats the character that comes after it enough times to fill
the column.

The asterisk is ignored in `cell.formattedValue`, and the character is
only repeated if the `formatValue` helper is given a `width` parameter,
which we now do only in the grid renderer.

The commit also adds this new format feature to the accounting format,
to match the behaviour of the other spreadsheet applications.

Task: 3965246
Part-of: odoo/o-spreadsheet#4961
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>