[IMP] composer: recolor parenthesis
Before this commit, targeting a parenthesis with the
cursor changed the color of the parenthesis and its
corresponding parenthesis.

Two changes have been made here:
- firstly the visual change from gray to black color of
the font color was not noticeable. We therefore preferred
here to change the color of the background to improve
perception.
- second, the change took place even when the parenthesis
did not have a corresponding  parenthesis. This issue has
been fixed.

We also added a red color to the excess parentheses in the
formula.

Task: 4123205