[IMP] function: better inferred format for MIN,MAX,MEDIAN
Currently the result format of MIN,MAX and MEDIAN function
is inferred from the very first cell of the given dataset.

If that cell has no format (because it's empty or any other
reason), then the result has no format.

Specifically for MAX and MIN, the format should be the format
of the max/min value.

For MEDIAN, the format is now inferred from the first numeric cell.

Task: 4137785
Part-of: odoo/o-spreadsheet#4904
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>