[FIX] Find&Replace: always search cell content if `searchFormula`
The current implementation of a search with the option `serachFormula`
will only look into the cell content if the cell is a valid formula cell
and not a cell which results in an error following its evaluation.

e.g.  A1: '=SUM("a")' is invalid thus the related cell is not a formula
cell but an error cell and will never be matched when looking for the
term "sum".

closes odoo/o-spreadsheet#4960

Task: 4175968
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>