[PERF] model: avoid useless finalize on initialization
Currently, finalize will be called after each replay of a
`stateUpdateMessage` during the Model instanciation. Depending on the
type of revisions and on the number of cells to evaluate, this can end
up having a devastating effect on the performance at first load.

2 examples:
- on a spreadsheet on 450*26 cells (all filled) with 300 random but
  simple revisions, we spend 2.8 seconds in the evaluaion
- A client spreadsheet in 16.0 with 1100 revisions cannot load due to a
  crash of the broweer tab

This revision adds a "loading state" where the finalize is disabled and
only used once after the replay of all stateUpdateMessages.

closes odoo/o-spreadsheet#4969

Task: 4176216
X-original-commit: 512617b9987aa138101dedf3ccc6d63cd584c6fa
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>