[FIX] pivot: running total display with multiple measures
The running total computation was wrong in pivot if there was multiple
measures: the pivot cells were not filtered by the measure on which
the running total was computed, which led to the same cell being
counted multiple times.

The display rank ascending/descending had theoretically the same issue,
but since pivot cells with the exact same value have the same rank, it
didn't end up impacting the final result.

closes odoo/o-spreadsheet#5020

Task: 4207244
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>