[FIX] ComposerStore: Allow to select the spill Pivot formula
Since #4445, Users can select a spilled pivot cell in the composer and
it will automatically be converted to its inidividual equivalent.
The behaviour was deactivated on purpose when the user would click on an
pivot formula but it should only target individual pivot formulas (e.g.
`PIVOT.VALUE`) and not spilled pivot formulas (i.e. `PIVOT`).

Task: 4207502