[IMP] cf: apply panel update immediately
The previous workflow of the CF side panel was to edit the conditional
format, then click on save to confirm. This wasn't the greatest, as it
meant that you wouldn't see the result of your changes until you saved.
And if you didn't like the change, you then needed to re-click on the
CF to re-open its editor, and do the changes.

With this commit, the CF side panel now updates immediately when you
change a value. This also mean that a new CF is created immediately
when you click on the "Add" button. The changes made can be discarded
by clicking on the "Cancel" button.

closes odoo/o-spreadsheet#5050

Task: 3720877
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>