[FIX] table: insert table without filters
When inserting a table, it is inserted with the filter. It takes place on the
cell and is only useful in a handful of use cases. It's very frequent that the
very next step after inserting a table is to remove it.

So let's insert the table without the filter by default.

Feedback from RNG

Task: 4235172
Part-of: odoo/o-spreadsheet#5076
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>