[FIX] collaborative: don't send duplicate messages
When a message is acknowledged, the `waitingAck` flag is reset to `false`,
but it's not set to `true` when the next pending message is sent.

It means that if a third local change occurs, it will send the next pending
message even though the second one is not acknowledged yet.

closes odoo/o-spreadsheet#5093

Task: 4255949
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>