[FIX] find and replace: do not update selection at each change
Previously, when doing any operation that would refresh the search
(eg. an update_cell), we would update the selection to the current
search result. This was preventing the user to have any work done
while the f&r panel was open.

This commit changes the behavior to only update the selection when
the user explicitly clicks on a search result or changes the sheet.

Task: 3973671
Part-of: odoo/o-spreadsheet#5101
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>