[FIX] chart: cleanup tooltip title code
The code that ensure that the the trend lines points in a chart have
no tooltip title is uselessly complex, with reference to some
"originalTooltipTitle" and a //@ts-expect-error.

In fact, originalTooltipTitle is almost always undefined, except for
linear charts where it's a simple callback that return an empty string.

This commit simplifies the code, removing the obscure reference to
"originalTooltipTitle" and the //@ts-expect-error. The callback now
simply returns an empty string where we do not want a title, and
undefined if we want to let ChartJs do it's thing.

closes odoo/o-spreadsheet#5103

Task: 4269632
X-original-commit: ab8b491c9e887c2193e532efc7119d40eee55b5a
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>