[REF] charts: simplify runtime generation
Currently to generate a chart runtime, every chart type does its own
thing, and the code is very spaghetti.

An example: the pie chart `createPieChartRuntime` function calls
`getPieConfiguration` that calls `getDefaultChartJsRuntime`. Every level
creates/modifies/overrides keys of the ChartJS configuration object,
and it's very hard to track what does what and what is the final result.

Or the bar chart does everything in a 150 lines function. And the
pyramid chart calls the bar chart config, then overrides some keys.
And I'm not even talking about odoo charts that copy/paste a bunch
of code from the spreadsheet charts for their runtime.

This commit is an attempt to make everything more readable and modular.
The idea is to separate the runtime generation in two steps:

1) extract the data from the spreadsheet.

The methods that does that have access to the getters and no concept
of ChartJS. They return an object with the data that will be used to
generate the ChartJS runtime.

2) generate the ChartJS runtime.

For each chartJS option (eg. the legend) we have a separate file that
contains helpers to generate the option for each chart type. This make
it easy to see what's the difference between the chart types.

The solution is not perfect, but it does make the code easier to follow.
And (hopefully) will allow us to re-use the runtime generation for the
odoo charts with all the features and no code duplication.

closes odoo/o-spreadsheet#5105

Task: 4286137
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>