[FIX] chart: trend line of area chart
The trend line of area charts should not have a filled area under
the line.

Also the trend line configuration should probably not inherit all the
values of their parent dataset, as it's easy to miss thing and have
options that make no sense (pointBackgroundColor being the color
of the parent dataset rather than the color of the trend line for
example).

closes odoo/o-spreadsheet#5182

Task: 4274294
X-original-commit: 2cb20cee617dc1835a2933d23dc886832939fe1b
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>