[FIX] logs: log performance information at debug level
3b1657f introduced the logging of performance information in the logs.
This could be useful in some case, but often flooded the console for
nothing.

This commit changes the log level to debug, so that it is not displayed
by default.

Unfortunately console.group is not meant for debug level (it still
group the messages, but the group title is always displayed even if
debug logs are not displayed). Replaced it by good'ol
`console.debug("#####")` to mark sections.

closes odoo/o-spreadsheet#5113

Task: 0
X-original-commit: 504b907365f8e58d58d4cd204bbfd7cbc33080ff
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>