[FIX] test: skipped test in `renderer_plugin.test.ts`
A test was skipped in `renderer_plugin.test.ts` because it didn't
contain code instructions, but it contained a callback with those
instructions. And the callback was never called.

The test was also wrong: it forgot to add `MIN_CELL_TEXT_MARGIN` to
the width of the text, and to resize the column to make the text
overflow.

closes odoo/o-spreadsheet#5121

Task: 4276968
X-original-commit: a36121f66ce0aba9e604bed58d3433a4857bcf51
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>