[FIX] functions: FILTER with strings and errors
The FILTER function should only consider numbers and booleans as the
conditions but it currently only accepts all types of values (strings, errors)
as long as the value is "truthy".

Steps to reproduce:

Let's consider this grid
	A	B
1     Alice  coucou
2      Bob    TRUE

The formula =FILTER(A1:A2, B1:B2) should only yield Bob and not Alice.

Note that this is a breaking change since the behavior is different.
That's why I'm not backporting the fix.

closes odoo/o-spreadsheet#5159

Task: 4307604
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>