[FIX] pivot: skip cells on manipulated pivot array formulas
Steps to reproduce in odoo:

- insert a pivot with more rows than columns
- write in a cell =TRANSPOSE(ODOO.PIVOT.TABLE(1))
- right-click on the grand total value

=> boom

When computing which cell of the pivot table is clicked, we assume the matrix
comes directly from the ODOO.PIVOT.TABLE(...) function to compute the offsets
from the array formula. But it's completely wrong as the cell could at a
completely different place if the matrix is manipulated by other functions
before being outputted to the grid.

closes odoo/o-spreadsheet#5162

X-original-commit: 360550579767b39ec8fc0fb2a07adc0363295127
Task: 4292134
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>