[REF] xlsx: remove some custom values in ExcelCellData
This commit prepares the ground before the `CellData` object is removed from
the export.

We must first untangle `ExcelCellData` from `CellData` and this commit
is a first step for this work.

When exporting the model data for excel, `ExcelCellData` is different from
`CellData`.
It has several additional values on the objects. Among those additional values
are `style`, `format` and `border`. This data is also present in `sheet.styles`,
`sheet.formats` and `sheet.borders` (see c3c0b45ed1eae582f5ad00)

At the time, it was easier to duplicate those values in the "augmented"
`ExcelCellData` as it allowed to keep the excel export unchanged.

This divergence has several issues:
- duplicating data is rarely a good idea (which one should I use? Are they
  always perfectly synced?)
- it's not as simple as it could be since there are 2 (almost identical) data
  schema instead of one.

This commit removes the additional `style`, `format` and `border` values and
the excel export now uses the same data exported by the regular export.

It makes the export code slightly more complicated, but simplifies the data
structures and `exportForExcel` implementation.

The next commit completely removes other additional values in `ExcelCellData`

Task: 4295344