[IMP] chart: improve bar chart borders
Task Description

Since the new design of chart color, when we try to draw a bar chart, we have
a small difference between a classical bar chart and a combo bar where every
series is a bar. This is due to the border of the bar in a bar chart that are
drawn in a whitish color, while the border are still in the same color as the
bar in a combo chart. This task aims to change the way we draw the border of
a bar chart, drawing borders only for stacked bar chart, in the same color of
the background of the chart.

Related Task

closes odoo/o-spreadsheet#5140

Task: 4283212
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>